Convert RoPE inv_freq buffer to bfloat16 to match model weights
Compute inv_freq in float32 (via automatic type promotion), then
convert to bfloat16 to match pretrained weights dtype.

Fixes RuntimeError: "expected m1 and m2 to have the same dtype,
but got: float != c10::BFloat16"

+++ rope.py
@@ -34,9 +34,10 @@ class RoPE(nn.Module):
         # Precompute the rotation frequencies
         # For each pair of dimensions, we have a different frequency
         # Shape: (head_dim // 2,)
-        # Always create this buffer on CPU since it's computed, not loaded from weights
+        # Compute in float32 for accuracy, then convert to bfloat16 to match model weights
         with torch.device("cpu"):
-            inv_freq = 1.0 / (self.theta ** (torch.arange(0, head_dim, 2).float() / head_dim))
+            inv_freq = 1.0 / (self.theta ** (torch.arange(0, head_dim, 2) / head_dim))
+            inv_freq = inv_freq.to(torch.bfloat16)
         self.register_buffer("inv_freq", inv_freq, persistent=False)
 
     def forward(self, x: torch.Tensor, position_ids: torch.Tensor) -> torch.Tensor:
